car_object_detection/requirements.txt
ultralytics
torch
pandas
Pillow
matplotlib
ultralytics
opencv-python
pandas
Pillow
fastapi
uvicorn
streamlit
matplotlib
